+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## Added
 
-- Component inheritance (controlled via a the new component level option: `inheritMethods`).
+- Component inheritance.
 - Return value from `noDispatch` methods.
 - Error logs.